+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ### Added
 
 - Support `java.util.Date` parcelization.
+- Support `java.util.UUID` and `kotlin.uuid.Uuid` parcelization.
 
 ## [0.1.0](https://github.com/Goooler/kaidl/releases/tag/0.1.0) - 2026-04-17